Finished def14.Perf.10 - api_performance_requests_per_second test. Closes #14
