--- ftp/tests/api_performance_requests_per_second/output.xml
@@ -1,150 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<robot generator="Robot 6.0.2 (Python 3.9.6 on darwin)" generated="20231228 12:07:15.721" rpa="false" schemaversion="3">
-<suite id="s1" name="Api Performance Requests Per Second" source="/Users/rdireito/Desktop/UA/5GASP/Code/CICD_LTR/ftp/tests/api_performance_requests_per_second">
-<suite id="s1-s1" name="Test" source="/Users/rdireito/Desktop/UA/5GASP/Code/CICD_LTR/ftp/tests/api_performance_requests_per_second/test.robot">
-<test id="s1-s1-t1" name="Run Locust Requests Per Second Test" line="7">
-<kw name="Run Process" library="Process">
-<var>${output}</var>
-<arg>locust</arg>
-<arg>-f</arg>
-<arg>test.py</arg>
-<arg>--headless</arg>
-<arg>-u</arg>
-<arg>50</arg>
-<arg>-r</arg>
-<arg>50</arg>
-<arg>--run-time</arg>
-<arg>5</arg>
-<arg>--host</arg>
-<arg>https://atnog-harbor.av.it.pt</arg>
-<arg>--endpoint</arg>
-<arg>/</arg>
-<arg>--json</arg>
-<doc>Runs a process and waits for it to complete.</doc>
-<msg timestamp="20231228 12:07:15.742" level="INFO">Starting process:
-locust -f test.py --headless -u 50 -r 50 --run-time 5 --host https://atnog-harbor.av.it.pt --endpoint / --json</msg>
-<msg timestamp="20231228 12:07:15.745" level="INFO">Waiting for process to complete.</msg>
-<msg timestamp="20231228 12:07:21.029" level="INFO">Process completed.</msg>
-<msg timestamp="20231228 12:07:21.029" level="INFO">${output} = &lt;result object with rc 0&gt;</msg>
-<status status="PASS" starttime="20231228 12:07:15.742" endtime="20231228 12:07:21.029"/>
-</kw>
-<kw name="Evaluate Locust Results" library="evaluate">
-<var>${median_requests_per_second}</var>
-<arg>${output.stdout}</arg>
-<msg timestamp="20231228 12:07:21.030" level="INFO">LOCUST RESULTS:
-[
-    {
-        "name": "/",
-        "method": "GET",
-        "last_request_timestamp": 1703765240.88841,
-        "start_time": 1703765236.5348501,
-        "num_requests": 137,
-        "num_none_requests": 0,
-        "num_failures": 0,
-        "total_response_time": 24906.134459000015,
-        "max_response_time": 447.26983299999995,
-        "min_response_time": 24.10237500000001,
-        "total_content_length": 132068,
-        "response_times": {
-            "380.0": 4,
-            "390.0": 7,
-            "410.0": 3,
-            "420.0": 4,
-            "430.0": 20,
-            "440.0": 6,
-            "450.0": 6,
-            "47": 4,
-            "46": 4,
-            "35": 3,
-            "55": 3,
-            "27": 2,
-            "33": 4,
-            "49": 3,
-            "24": 1,
-            "30": 1,
-            "63": 1,
-            "52": 3,
-            "59": 3,
-            "50": 3,
-            "43": 4,
-            "44": 2,
-            "32": 3,
-            "34": 3,
-            "40": 2,
-            "58": 1,
-            "39": 6,
-            "42": 6,
-            "48": 4,
-            "87": 1,
-            "36": 1,
-            "61": 1,
-            "60": 1,
-            "45": 4,
-            "25": 1,
-            "29": 1,
-            "51": 3,
-            "53": 1,
-            "41": 3,
-            "62": 1,
-            "38": 1,
-            "37": 1,
-            "64": 1
-        },
-        "num_reqs_per_sec": {
-            "1703765236": 50,
-            "1703765237": 12,
-            "1703765238": 27,
-            "1703765239": 23,
-            "1703765240": 25
-        },
-        "num_fail_per_sec": {}
-    }
-]
-Test Duration: 5 seconds.
-Mean Requests per Second: 27.4
-Median Requests per Second: 25</msg>
-<msg timestamp="20231228 12:07:21.030" level="INFO">${median_requests_per_second} = 25</msg>
-<status status="PASS" starttime="20231228 12:07:21.029" endtime="20231228 12:07:21.030"/>
-</kw>
-<if>
-<branch type="IF" condition="${median_requests_per_second} &gt;= %{api_performance_requests_per_second_min_threshold}">
-<kw name="Pass Execution" library="BuiltIn">
-<arg>\nThe Target Server/API COMPLIES with the minimum threshold set for number of requests/sec (%{api_performance_requests_per_second_min_threshold}).</arg>
-<doc>Skips rest of the current test, setup, or teardown with PASS status.</doc>
-<status status="NOT RUN" starttime="20231228 12:07:21.031" endtime="20231228 12:07:21.031"/>
-</kw>
-<status status="NOT RUN" starttime="20231228 12:07:21.030" endtime="20231228 12:07:21.031"/>
-</branch>
-<branch type="ELSE">
-<kw name="Fail" library="BuiltIn">
-<arg>\nThe Target Server/API DOES NOT comply with the minimum threshold set for number of requests/sec (%{api_performance_requests_per_second_min_threshold}).</arg>
-<doc>Fails the test with the given message and optionally alters its tags.</doc>
-<msg timestamp="20231228 12:07:21.031" level="FAIL">
-The Target Server/API DOES NOT comply with the minimum threshold set for number of requests/sec (50).</msg>
-<status status="FAIL" starttime="20231228 12:07:21.031" endtime="20231228 12:07:21.031"/>
-</kw>
-<status status="FAIL" starttime="20231228 12:07:21.031" endtime="20231228 12:07:21.031"/>
-</branch>
-<status status="FAIL" starttime="20231228 12:07:21.030" endtime="20231228 12:07:21.031"/>
-</if>
-<status status="FAIL" starttime="20231228 12:07:15.742" endtime="20231228 12:07:21.031">
-The Target Server/API DOES NOT comply with the minimum threshold set for number of requests/sec (50).</status>
-</test>
-<status status="FAIL" starttime="20231228 12:07:15.733" endtime="20231228 12:07:21.031"/>
-</suite>
-<status status="FAIL" starttime="20231228 12:07:15.723" endtime="20231228 12:07:21.032"/>
-</suite>
-<statistics>
-<total>
-<stat pass="0" fail="1" skip="0">All Tests</stat>
-</total>
-<tag>
-</tag>
-<suite>
-<stat pass="0" fail="1" skip="0" id="s1" name="Api Performance Requests Per Second">Api Performance Requests Per Second</stat>
-<stat pass="0" fail="1" skip="0" id="s1-s1" name="Test">Api Performance Requests Per Second.Test</stat>
-</suite>
-</statistics>
-<errors>
-</errors>
-</robot>
